+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2023.3.4]
+
+### `Project Build` enhancements
+- Make code generation timeout configurable [#954](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/954)
+
 ## [2023.3.3]
 
 ### Features